fix: add ignore rule for PowerToys Command Palette to sample config (#1140)
Microsoft recently added a new Command Palette/Runner to PowerToys and
it's getting tiled ( #1138 ).

After fighting with WinSpy++ I think I found the correct title + class
combination to ignore it. Works for me, but I'm still new to both the
project and "Windows windows inspections". If I messed up something
else, or maybe there's a more optimal way to do this change, lmk (or
take inspiration and close this commit - I'm ok with that).

Thanks for software and have a great day!